+++ CHANGELOG.md
@@ -17,3 +17,4 @@
 
 ## 2026-03-11
 - docs: end-of-day summary at 2026-03-11 08:21:32
+- feat: add workspace project structure validator with RED-GREEN test coverage